Add message class
